Moved progress dialog to appropriate module
