+++ app/build.gradle
@@ -53,6 +53,8 @@ android {
 
 dependencies {
 
+    implementation project(':foundation:dialog')
+
     implementation deps.app.androidx.appcompat
     implementation deps.app.androidx.core_ktx
     implementation deps.app.androidx.core_splashscreen
