+++ foundation/dialog/build.gradle
@@ -1,0 +1,45 @@
+plugins {
+    id 'com.android.library'
+    id 'org.jetbrains.kotlin.android'
+    id 'kotlin-kapt'
+    id 'com.google.dagger.hilt.android'
+    id 'org.jlleitschuh.gradle.ktlint'
+}
+
+android {
+    namespace 'com.weather.foundation.dialog'
+    compileSdk rootProject.compileSdkVersion
+
+    defaultConfig {
+        applicationId "com.weather.foundation.dialog"
+        minSdk rootProject.minSdkVersion
+        targetSdk rootProject.targetSdkVersion
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
+    }
+    compileOptions {
+        sourceCompatibility environment.java_version
+        targetCompatibility environment.java_version
+    }
+    kotlinOptions {
+        jvmTarget = environment.java_version
+    }
+}
+
+dependencies {
+
+    implementation project(':foundation:resources')
+
+    implementation deps.app.androidx.appcompat
+    implementation deps.app.androidx.core_ktx
+
+    implementation deps.app.google.hilt_android
+    kapt deps.app.google.hilt_compiler
+
+    implementation deps.app.material
+}
