+++ foundation/dialog/src/main/res/layout/dialog_progress.xml
