+++ settings.gradle
@@ -14,3 +14,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "Weather"
 include ':app'
+include ':foundation:dialog'
